Allow user to create persistent context and queue
Having context and queue span over multiple kernel calls would allow us
to take advantage of the OpenCL scheduler. This also enables us to offer
OpenCL buffers directly to the user, so that he can keep his data on the
GPU in between kernel calls. And it's not unthinkable that this brings a
performance benefit.

The downside is that the user has to call one additional, but we might
mitigate this in the future by offering a sensible default for the
device parameter.