Update documentation of oclSimpleKernel and oclRun
The changes reflect the introduction of OpenCL buffers into oclRun.
Additionally, the flags native.result and wait have been removed, and
the method oclResult to collect the results of asynchronous calls. The
first flag is no longer needed, instead users have to convert their
vectors to single-precision buffers manually. The second will hopefully
be replaced by a superior mechanism soon that attaches events to buffers
and makes use of out-of-order execution in OpenCL command queues.