Expose single-precision buffers as numeric vectors
Reading from single-precision buffers now yields a numeric vector
instead of a clFloat object. Also, writing to such a buffer now
requires a numeric vector.

The mode names have been changed to "single" and "double", which is
pretty close to R-terminology. The mode "numeric" remains as an alias to
"double" (for now).