Remove host-side single-precision vector type
This is no longer needed, because the conversion happens when we read
from and write to buffers. That makes more sense than keeping single-
precision buffers on the host side, where we can't use them anyway.

The bottleneck will also be more likely the transfer between CPU and
GPU, while the conversion is rather fast.