Print build log in its full glory
It turns out that if the buffer passed into clGetProgramBuildInfo is not
large enough, we don't get a single byte of the build log. So we need to
use an apparently undocumented feature to obtain the build log length,
then allocate a buffer for the log dynamically.

Also adds more error handling to clGetProgramBuildInfo, and fixes a
minor issue in the oclRun documentation example.