Always write prototypes for functions
An empty parameter list is not a function prototype in C (unlike C++).
Calling a non-prototype (aka K&R declaration) with the wrong number of
parameters is not a guaranteed compiler error, so these should be
avoided. (In fact K&R declarations are considered deprecated since C89.)

GCC and Clang can find non-prototype functions via -Wstrict-prototypes.